Corrigindo Address Observer para passar nos testes.

+++ app/Domains/Ad/Observers/AddressObserver.php
@@ -36,6 +36,8 @@ class AddressObserver
      */
     public function deleted(Ad $ad)
     {
-        $ad->address->delete();
+        if ($ad->address) {
+            $ad->address->delete();
+        }
     }
 }
